JSX to TSX conversion from Company job listing page to Premium
