+++ src/Screens/Edit Company Page/Edit_Company_Page.tsx
@@ -1,13 +1,15 @@
-import { React, Component } from "react";
+import React = require("react")
+import { Component } from "react";
 import "../index.css";
 import "../css/home.css";
 import "../css/login.css";
 import "bootstrap/dist/css/bootstrap.min.css";
-import Navbar_company from "./Navbar_Company";
+import Navbar_company from "../Navbar/Navbar_Company";
 import { Form, Input } from "antd";
 import axios from "axios";
 import FormItem from "antd/lib/form/FormItem";
 class editcompany extends Component {
+  state: { title: string, location: string, company_description: string, id: string, website: string, facebook: string, twitter: string, linkedin: string, instagram: string, vedio_link: string }
   constructor(props) {
     var arr = JSON.parse(localStorage.getItem("company"));
     super(props);
@@ -90,7 +92,7 @@ class editcompany extends Component {
       Connection: "keep-alive",
     };
     axios
-      .post("https://camprec.herokuapp.com/api/company/edit", body, { header })
+      .post("https://camprec.herokuapp.com/api/company/edit", body, { headers : {header} })
       .then(function (response) {
         console.log("*");
         const data = response.data;
