+++ src/Screens/Navbar/Navbar_Show_Applicants.tsx
@@ -2,16 +2,16 @@ import React from 'react';
 import { NavLink } from "react-router-dom";
 import '../index.css';
 import '../css/navbar.css'
-import web from "../images/Capture-removebg-preview.png";
+const web = require("../images/Capture-removebg-preview.png");
 
-const company_Navbar = () => {
+const viewcollege_Navbar = () => {
     return (
         <>
        <div className = "container-fluid nav_bg">
               <div className = "row navbar">
                   <div className = "col-10 mx-auto navbar">
 
-                  <nav class="navbar navbar-expand-lg navbar-light">
+                  <nav className="navbar navbar-expand-lg navbar-light">
   <div className="container-fluid navbar">
     <NavLink  className="navbar-brand" to = "/">
     <img src={web} className = "image-fluid  logo padl" alt = "home image"/>
@@ -34,27 +34,7 @@ const company_Navbar = () => {
           <NavLink
           activeClassName = "menu_active" 
           className="nav-link" 
-          to = "/companyjobs"
-          >
-            jobs
-          
-          </NavLink>
-        </li>
-        <li className="nav-item">
-          <NavLink
-          activeClassName = "menu_active" 
-          className="nav-link" 
-          to = "/companycolleges"
-          >
-            Colleges
-          
-          </NavLink>
-        </li>
-        <li className="nav-item">
-          <NavLink
-          activeClassName = "menu_active" 
-          className="nav-link" 
-          to = "/logout">Logout</NavLink>
+          to = "/companyjobs">Back</NavLink>
           </li>
       </ul>
     </div>
@@ -67,4 +47,4 @@ const company_Navbar = () => {
     );
 };
 
-export default company_Navbar;
+export default viewcollege_Navbar;
