+++ src/Screens/Navbar/Navbar_View_Applicant.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { NavLink } from "react-router-dom";
 import '../index.css';
 import '../css/navbar.css'
-import web from "../images/Capture-removebg-preview.png";
+const web = require("../images/Capture-removebg-preview.png");
 
 const viewcollege_Navbar = () => {
     return (
@@ -11,7 +11,7 @@ const viewcollege_Navbar = () => {
               <div className = "row navbar">
                   <div className = "col-10 mx-auto navbar">
 
-                  <nav class="navbar navbar-expand-lg navbar-light">
+                  <nav className="navbar navbar-expand-lg navbar-light">
   <div className="container-fluid navbar">
     <NavLink  className="navbar-brand" to = "/">
     <img src={web} className = "image-fluid  logo padl" alt = "home image"/>
