+++ src/Screens/Navbar/Navbar_View_College.tsx
@@ -2,16 +2,17 @@ import React from "react";
 import { NavLink } from "react-router-dom";
 import "../index.css";
 import "../css/navbar.css";
-import web from "../images/Capture-removebg-preview.png";
-const Navbar_viewstdcompany = () => {
+const web = require("../images/Capture-removebg-preview.png");
+
+const viewcollege_Navbar = () => {
   return (
     <>
       <div className="container-fluid nav_bg">
         <div className="row navbar">
           <div className="col-10 mx-auto navbar">
-            <nav class="navbar navbar-expand-lg navbar-light">
+            <nav className="navbar navbar-expand-lg navbar-light">
               <div className="container-fluid navbar">
-                <NavLink className="navbar-brand" to="/studenthome">
+                <NavLink className="navbar-brand" to="/">
                   <img
                     src={web}
                     className="image-fluid  logo padl"
@@ -29,36 +30,28 @@ const Navbar_viewstdcompany = () => {
                         exact
                         className="nav-link active"
                         aria-current="page"
-                        to="/studenthome"
+                        to="/viewcompanycollege"
                       >
                         Home
                       </NavLink>
                     </li>
+
                     <li className="nav-item">
                       <NavLink
                         activeClassName="menu_active"
                         className="nav-link"
-                        to="/studentjobs"
+                        to="/viewcollegestudent"
                       >
-                        jobs
+                        Students
                       </NavLink>
                     </li>
                     <li className="nav-item">
                       <NavLink
                         activeClassName="menu_active"
                         className="nav-link"
-                        to="/viewstudentcompany"
+                        to="/companycolleges"
                       >
-                        back
-                      </NavLink>
-                    </li>
-                    <li className="nav-item">
-                      <NavLink
-                        activeClassName="menu_active"
-                        className="nav-link"
-                        to="/logout"
-                      >
-                        Logout
+                        Back
                       </NavLink>
                     </li>
                   </ul>
@@ -72,4 +65,4 @@ const Navbar_viewstdcompany = () => {
   );
 };
 
-export default Navbar_viewstdcompany;
+export default viewcollege_Navbar;
