+++ src/Screens/Navbar/Navbar_View_Std_Company_2.tsx
@@ -2,17 +2,16 @@ import React from "react";
 import { NavLink } from "react-router-dom";
 import "../index.css";
 import "../css/navbar.css";
-import web from "../images/Capture-removebg-preview.png";
-
-const viewcollege_Navbar = () => {
+const web = require("../images/Capture-removebg-preview.png");
+const Navbar_viewstdcompany = () => {
   return (
     <>
       <div className="container-fluid nav_bg">
         <div className="row navbar">
           <div className="col-10 mx-auto navbar">
-            <nav class="navbar navbar-expand-lg navbar-light">
+            <nav className="navbar navbar-expand-lg navbar-light">
               <div className="container-fluid navbar">
-                <NavLink className="navbar-brand" to="/">
+                <NavLink className="navbar-brand" to="/studenthome">
                   <img
                     src={web}
                     className="image-fluid  logo padl"
@@ -30,28 +29,36 @@ const viewcollege_Navbar = () => {
                         exact
                         className="nav-link active"
                         aria-current="page"
-                        to="/viewcompanycollege"
+                        to="/studenthome"
                       >
                         Home
                       </NavLink>
                     </li>
-
                     <li className="nav-item">
                       <NavLink
                         activeClassName="menu_active"
                         className="nav-link"
-                        to="/viewcollegestudent"
+                        to="/studentjobs"
                       >
-                        Students
+                        jobs
                       </NavLink>
                     </li>
                     <li className="nav-item">
                       <NavLink
                         activeClassName="menu_active"
                         className="nav-link"
-                        to="/companycolleges"
+                        to="/viewstudentcompany"
                       >
-                        Back
+                        back
+                      </NavLink>
+                    </li>
+                    <li className="nav-item">
+                      <NavLink
+                        activeClassName="menu_active"
+                        className="nav-link"
+                        to="/logout"
+                      >
+                        Logout
                       </NavLink>
                     </li>
                   </ul>
@@ -65,4 +72,4 @@ const viewcollege_Navbar = () => {
   );
 };
 
-export default viewcollege_Navbar;
+export default Navbar_viewstdcompany;
