+++ tsconfig.json
@@ -1,0 +1,7 @@
+{
+    "compilerOptions" : {
+        "jsx": "react",
+        "esModuleInterop" : true,
+    }
+
+}
